feat(web): the api/ stubs leave; the Build Output tree is the deployment
Vercel's zero-config pass builds every file under api/ beside the Build
Output tree the build now writes, two builders claiming one path, so the
eight generated one-line stubs go and the feedback function's source moves
to server/feedback.mjs, emitted into the tree at its unchanged URL. The two
under api/voice/ were re-exports of their bundles and nothing else:
server/voice/* and server/routes/voice/* do not move.

The stubs existed because Vercel registered api/ functions from the uploaded
tree before the build ran (#1140's function-stubs.ts says so); with the tree
the build's functions deploy directly and that reason inverts. What retires
with them is the stub half of that module's check and only that half:
function-stubs.ts becomes function-layout.ts, keeping the route keys, the
function definitions, the bundle paths, and the public path (the former
stubPath, still api/<function>.js, the .func name and the rewrites'
destination); scripts/function-stubs.ts becomes scripts/function-rewrites.ts,
whose --write regenerates vercel.json's /api/ rewrites and whose --check
fails when they drift, unchanged; repository-checks.sh runs that check and no
longer greps a vanished api/ for Effect layers. The rewrites-drift guarantee
is the same guarantee: committed, generated, checked.
