--- apps/web/api/brain-embed.js
@@ -1,2 +1,0 @@
-export { default } from "../dist-functions/brain-embed.js";
-export const config = { maxDuration: 60 };
