--- apps/web/api/brain-inference.js
@@ -1,2 +1,0 @@
-export { default } from "../dist-functions/brain-inference.js";
-export const config = { maxDuration: 120 };
