--- apps/web/api/default.js
@@ -1,1 +1,0 @@
-export { default } from "../dist-functions/default.js";
