--- apps/web/api/observation-tick.js
@@ -1,2 +1,0 @@
-export { default } from "../dist-functions/observation-tick.js";
-export const config = { maxDuration: 60 };
