--- apps/web/api/turn-events.js
@@ -1,2 +1,0 @@
-export { default } from "../dist-functions/turn-events.js";
-export const config = { maxDuration: 300 };
