--- apps/web/api/turn-read.js
@@ -1,2 +1,0 @@
-export { default } from "../dist-functions/turn-read.js";
-export const config = { maxDuration: 40 };
