--- apps/web/api/voice/introduction.js
@@ -1,2 +1,0 @@
-export { default } from "../../dist-functions/voice/introduction.js";
-export const config = { maxDuration: 800 };
