--- apps/web/api/voice/sessions.js
@@ -1,2 +1,0 @@
-export { default } from "../../dist-functions/voice/sessions.js";
-export const config = { maxDuration: 800 };
