+++ apps/web/server/README.md
@@ -81,21 +81,23 @@ Google's callback is `${BETTER_AUTH_URL}/api/auth/callback/google`; GitHub's is
 `user:email`, because Luke requires an email address for its account snapshot.
 
 Every function Vercel deploys is plain ESM. The route sources live under
-`server/routes/`, mirroring the `api/` tree, and `scripts/bundle-functions.ts`
-bundles them into `dist-functions/**/*.js` (gitignored) as the last step of
-`pnpm build`, each bundle inlined whole: the workspace packages and every
-declared runtime dependency alike, so a function carries what it loads and
-nothing is left for a tracer to find. Handed TypeScript, the builder compiled
-every function's whole import graph separately, and those thirty-odd passes
-were most of a deploy's build time. What `api/` holds is a committed one-line
-stub per route, generated by `pnpm functions:stubs`, that re-exports its bundle
-and carries the route's `config` literal: Vercel registers `api/` functions
-from the uploaded tree before the build runs, so a function first written by
-the build is never deployed, and the builder reads `config` with a static
-parser that follows no re-export. Run `functions:stubs` after adding, moving,
-or removing a route; the bundle step and `repository-checks.sh` both refuse a
-stub that is missing, stale, or has no route behind it. `api/feedback.mjs` is
-the one hand-written function and stays plain ESM for the same reason.
+`server/routes/`, and `scripts/bundle-functions.ts` bundles them into
+`dist-functions/**/*.js` (gitignored) as the last step of `pnpm build`, each
+bundle inlined whole: the workspace packages and every declared runtime
+dependency alike, so a function carries what it loads and nothing is left for
+a tracer to find. Handed TypeScript, the builder compiled every function's
+whole import graph separately, and those thirty-odd passes were most of a
+deploy's build time. Nothing is committed under `api/`: the functions Vercel
+deploys are the ones the Build Output tree below emits, and a file under
+`api/` would be built by Vercel's zero-config pass beside that tree, two
+builders claiming one path. A function's public path is still
+`api/<function>.js` (`functionPublicPath` in `server/function-layout.ts`),
+because that is the `.func` name in the tree and the destination the `/api/`
+rewrites of `vercel.json` carry; `pnpm functions:rewrites` regenerates those
+rewrites after adding, moving, or removing a route, and
+`repository-checks.sh` refuses stale ones. `server/feedback.mjs` is the one
+hand-written function, emitted into the tree at `api/feedback.mjs` as it
+stands.
 
 Three specifiers stay outside every bundle, named in
 `INLINE_EXCEPTION`: `pg-native`, which `pg` tries before its JavaScript
@@ -162,8 +164,9 @@ identify. `docs/adr/0001-effect.md` names this edge with the process's others.
 The layer carries what a Vercel function's own platform already offers: an
 `HttpClient` over `fetch`, and the `SqlClient` of `server/db/sql-client.ts` over
 the database `DATABASE_URL` names. `@effect/platform-node` is not on it, and
-`repository-checks.sh` refuses that specifier and `@effect/sql*` under `api/`,
-where the stubs that re-export a bundle stand; the layer itself is behind
+`repository-checks.sh` refuses that specifier and `@effect/sql*` in the
+renderer, where there is no Node; a function inlines its whole graph and the
+isolation guard proves it loads. The layer itself is behind
 `server/`, which is what the builder traces. A function is handed no shutdown
 hook — an instance is frozen between invocations and discarded without notice —
 so nothing in production disposes the runtime; `disposeWebRuntime()` exists so a
