+++ knip.json
@@ -42,15 +42,14 @@
     "apps/web": {
       "entry": [
         "server/routes/**/*.ts",
-        "api/**/*.js",
-        "api/**/*.mjs",
         "src/*.tsx",
         "src/entry-server.tsx",
         "server/db/migrate.ts",
         "scripts/*.ts",
-        "tests/**/*.test.ts"
+        "tests/**/*.test.ts",
+        "server/feedback.mjs"
       ],
-      "project": ["api/**", "src/**", "server/**", "scripts/**", "tests/**"],
+      "project": ["src/**", "server/**", "scripts/**", "tests/**"],
       "ignoreDependencies": ["@sidecar/providers"]
     },
     "packages/*": {
